Moved sorting into separate library
