+++ src/multithreading/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_library(multithreading
         SpinningLock.cpp
-        MultithreadingUtils.cpp
+        Multithreading.cpp
 )
 
 target_include_directories(multithreading PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
